Server/requirements.txt
# Core Framework
Flask
Werkzeug
python-dotenv

# Database & ORM
Flask-SQLAlchemy
SQLAlchemy
Flask-Migrate
psycopg2-binary

# Authentication & Security
Flask-JWT-Extended
Flask-Cors
email-validator

# Async Tasks & Caching & Realtime
Celery
redis
Flask-Caching
Flask-SocketIO
simple-websocket

# Validation & Serialization
Flask-Marshmallow
marshmallow-sqlalchemy
marshmallow

# Production Server
gunicorn
gevent

# Git Integration (Phase 3 Prep)
GitPython

# AI/Agent Integration (Phase 6 Prep)
requests

# Testing & Quality
pytest
pytest-flask
pytest-mock
black
flake8
